+++ client/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added Works Caused By or Influenced By to Event pages ([#324](https://github.com/project-lux/lux-frontend/issues/324)).
+
 ### Changed
 
 - Changed mobile layout of place page ([#432](https://github.com/project-lux/lux-frontend/issues/432)).